+++ src/components/productContent/productWidget.tsx
