+++ src/pages/contactUs.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/space-before-function-paren */
 import React from 'react'
-import NavBar from '../components/navBar'
-import Footer from '../components/footer'
+import NavBar from '../components/navBar/navBar'
+import Footer from '../components/footer/footer'
 
 function ContactUs(): React.ReactElement {
   return (
