+++ src/pages/landingPage.tsx
@@ -1,15 +1,15 @@
 import React from 'react'
-import Navbar from '../components/navBar'
-import ProductContent from '../components/productContent'
-import BannerWrapper from '../components/bannerWrapper'
-import ProductWidget from '../components/productWidget'
-import ShippingDescWrapper from '../components/shippingDescWrapper'
-import SubscribeWrapper from '../components/subscribeWrapper'
-import CategoryWrapper from '../components/categoryWrapper'
-import Footer from '../components/footer'
-import FeaturedProductWrapper from '../components/featuredProductWrapper'
-import HeroContent from '../components/heroContent'
-import ReviewWrapper from '../components/reviewWrapper'
+import Navbar from '../components/navBar/navBar'
+import ProductContent from '../components/productContent/productContent'
+import BannerWrapper from '../components/Wrapper/bannerWrapper'
+import ProductWidget from '../components/productContent/productWidget'
+import ShippingDescWrapper from '../components/Wrapper/shippingDescWrapper'
+import SubscribeWrapper from '../components/Wrapper/subscribeWrapper'
+import CategoryWrapper from '../components/Wrapper/categoryWrapper'
+import Footer from '../components/footer/footer'
+import FeaturedProductWrapper from '../components/Wrapper/featuredProductWrapper'
+import HeroContent from '../components/productContent/heroContent'
+import ReviewWrapper from '../components/Wrapper/reviewWrapper'
 
 const landingPage = (): React.ReactElement => {
   return (
